refactor(google): Move GoogleAPIException to top level
Particularly as this starts having subclasses for particular
exception types, it makes sense for it to be at the top level,
not nested in the GoogleCommonSafeRetry.
