+++ clouddriver-appengine/src/main/groovy/com/netflix/spinnaker/clouddriver/appengine/deploy/AppengineSafeRetry.java
@@ -20,8 +20,8 @@ import com.google.common.collect.ImmutableList;
 import com.netflix.spectator.api.Registry;
 import com.netflix.spinnaker.clouddriver.appengine.deploy.exception.AppengineOperationException;
 import com.netflix.spinnaker.clouddriver.data.task.Task;
+import com.netflix.spinnaker.clouddriver.googlecommon.deploy.GoogleApiException;
 import com.netflix.spinnaker.clouddriver.googlecommon.deploy.GoogleCommonSafeRetry;
-import com.netflix.spinnaker.clouddriver.googlecommon.deploy.GoogleCommonSafeRetry.GoogleApiException;
 import com.netflix.spinnaker.kork.annotations.NonnullByDefault;
 import groovy.lang.Closure;
 import java.util.List;
