+++ clouddriver-google/src/main/groovy/com/netflix/spinnaker/clouddriver/google/deploy/SafeRetry.java
@@ -19,9 +19,9 @@ package com.netflix.spinnaker.clouddriver.google.deploy;
 import com.netflix.spectator.api.Registry;
 import com.netflix.spinnaker.clouddriver.data.task.Task;
 import com.netflix.spinnaker.clouddriver.google.deploy.exception.GoogleOperationException;
+import com.netflix.spinnaker.clouddriver.googlecommon.deploy.GoogleApiException;
+import com.netflix.spinnaker.clouddriver.googlecommon.deploy.GoogleApiException.ResourceInUseException;
 import com.netflix.spinnaker.clouddriver.googlecommon.deploy.GoogleCommonSafeRetry;
-import com.netflix.spinnaker.clouddriver.googlecommon.deploy.GoogleCommonSafeRetry.GoogleApiException;
-import com.netflix.spinnaker.clouddriver.googlecommon.deploy.GoogleCommonSafeRetry.ResourceInUseException;
 import com.netflix.spinnaker.kork.annotations.NonnullByDefault;
 import groovy.lang.Closure;
 import java.util.List;
